Move WERTracker
It might be useful for other recipes.
